ENG-4833: herald integration adding dictionary
