+++ packages/nepal-herald-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@al/herald",
-  "version": "2.0.8",
+  "version": "2.0.9",
   "license": "MIT",
   "description": "A lightweight client library for interacting with the Herald API",
   "author": {
